+++ grunt-tasks/tasks/nodeunit.js
@@ -1,7 +1,7 @@
 module.exports = function(grunt, options) {
 	return {
 		options: {
-			reporter: 'nested'
+			reporter: 'tap'
 		},
 		'veams-export': [ '<%= paths.tests %>/test-*.js' ]
 	};